# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todostock API",
        "description": "Documentación oficial de los endpoints de la API de Todostock para Todotek S.A. Incluye autenticación por Sanctum, gestión de inventarios y facturación transaccional.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor Local de Desarrollo (Docker Sail)"
        }
    ],
    "paths": {
...
        "/clients": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener listado de clientes paginado",
                "description": "Retorna una lista pagipada de los clientes registrados. Permite buscar de forma dinámica por nombre, identificación o email utilizando el parámetro 'search'.",
                "operationId": "getClientsList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda para filtrar clientes (nombre, identificación, email)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página para la paginación del servidor",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de clientes obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "identification": {
                                                        "type": "string",
                                                        "example": "0987654321"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "juan.perez@example.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "0999999999"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Av. Principal y Calle 2"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-05-19T00:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-05-19T00:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/clients?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/clients?page=3"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/clients?page=2"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost/api/clients"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 30
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado (Token inválido o ausente)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Registrar un nuevo cliente",
                "description": "Crea un registro de cliente en el sistema. Valida campos únicos como identificación y correo electrónico.",
                "operationId": "storeClient",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identification",
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "identification": {
                                        "type": "string",
                                        "example": "0955555555"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Carlos Mendoza"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos.mendoza@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0988888888"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Cdla. El Recreo, Durán"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 31
                                        },
                                        "identification": {
                                            "type": "string",
                                            "example": "0955555555"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Carlos Mendoza"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "carlos.mendoza@example.com"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "0988888888"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "Cdla. El Recreo, Durán"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-19T00:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-19T00:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación (Campos inválidos o duplicados)"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token 'access_token' recibido en el login con el formato: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}